+++ pkg/ml/layers/fnn/fnn.go
@@ -69,6 +69,8 @@ type Config struct {
 	input                           *Node
 	outputDimensions                []int
 	numHiddenLayers, numHiddenNodes int
+	ensembleSize                    int
+	ensembleAxis                    int
 	activation                      activations.Type
 	normalization                   string
 	dropoutRatio                    float64
@@ -117,6 +119,7 @@ func New(ctx *context.Context, input *Node, outputDimensions ...int) *Config {
 		ctx:              ctx,
 		input:            input,
 		outputDimensions: outputDimensions,
+		ensembleAxis:     -1,
 		numHiddenLayers:  context.GetParamOr(ctx, ParamNumHiddenLayers, 0),
 		numHiddenNodes:   context.GetParamOr(ctx, ParamNumHiddenNodes, 10),
 		activation:       activations.FromName(context.GetParamOr(ctx, activations.ParamActivation, "relu")),
@@ -134,6 +137,40 @@ func New(ctx *context.Context, input *Node, outputDimensions ...int) *Config {
 	if c.dropoutRatio < 0 {
 		c.dropoutRatio = context.GetParamOr(ctx, layers.ParamDropoutRate, 0.0)
 	}
+	return c
+}
+
+// WithEnsembleSize configure an ensemble size greater than 1, which adds an extra "ensemble axis"
+// to the variables and intermediary layers, executing the FNN as an ensemble.
+//
+// See WithEnsembleAxis for an alternative way to configure ensembles, when your input is already
+// split into the ensemble.
+func (c *Config) WithEnsembleSize(ensembleSize int) *Config {
+	c.ensembleSize = ensembleSize
+	return c
+}
+
+// WithEnsembleAxis defined an axis from the input operand that should be
+// considered as separate per model of the ensemble.
+//
+// It initializes the ensemble size from the input shape's given axis.
+//
+// This is used when constructing upper layers of an ensemble model, when the input has
+// already been transformer be per model of the ensemble -- hence it already has an ensemble axis.
+//
+// See WithEnsembleSize if you want to configure an ensemble for a plain input, which hasn't been
+// transformed per model of the ensemble, and hence doesn't have an ensemble axis.
+func (c *Config) WithEnsembleAxis(ensembleAxis int) *Config {
+	if ensembleAxis < 0 {
+		ensembleAxis = MustAdjustAxis(ensembleAxis, c.input)
+	}
+	c.ensembleAxis = ensembleAxis
+	dim := c.input.Shape().Dimensions[ensembleAxis]
+	if c.ensembleSize > 0 && c.ensembleSize != dim {
+		exceptions.Panicf("fnn: WithEnsembleAxis(%d), input shape is %s, corresponding dimension is %d, but ensembleSize is already %d",
+			ensembleAxis, c.input.Shape(), dim, c.ensembleSize)
+	}
+	c.ensembleSize = dim
 	return c
 }
 
@@ -213,8 +250,8 @@ func (c *Config) Regularizer(regularizer regularizers.Regularizer) *Config {
 // The default is 0.0, but it can be overridden by setting the hyperparameter layers.ParamDropoutRate (="dropout_rate")
 // in the context.
 func (c *Config) Dropout(ratio float64) *Config {
-	if ratio < 0 || ratio >= 1.0 {
-		exceptions.Panicf("fnn: invalid dropout ratio %f -- set to 0.0 to disable it, and it must be < 1.0 otherwise everything is dropped out",
+	if ratio >= 1.0 {
+		exceptions.Panicf("fnn: invalid dropout ratio %f -- set to <= 0.0 to disable it, and it must be < 1.0 otherwise everything is dropped out",
 			ratio)
 	}
 	c.dropoutRatio = ratio
@@ -227,6 +264,8 @@ func (c *Config) Done() *Node {
 	x := c.input
 	g := x.Graph()
 	dtype := x.DType()
+
+	isEnsemble := c.ensembleSize > 1
 
 	var dropoutRatio *Node
 	if c.dropoutRatio > 0.0 {
@@ -266,13 +305,31 @@ func (c *Config) Done() *Node {
 				x = Add(x, residual)
 			}
 			residual = x
+
+			// Adjust residual shape alignment to match the structural shape
+			// transition executed by DotGeneral for layers passing from ii=1 to ii>1
+			if isEnsemble && c.ensembleAxis < 0 && ii == 1 {
+				perm := make([]int, residual.Rank())
+				batchRank := residual.Rank() - 1 - len(outputDims)
+				perm[0] = batchRank
+				for j := 0; j < batchRank; j++ {
+					perm[j+1] = j
+				}
+				for j := batchRank + 1; j < residual.Rank(); j++ {
+					perm[j] = j
+				}
+				residual = TransposeAllAxes(residual, perm...)
+			}
 		}
 
 		// Linear transformation
 		inputLastDimension := x.Shape().Dimensions[x.Rank()-1]
-		weightsDims := make([]int, 1+len(outputDims))
-		weightsDims[0] = inputLastDimension
-		copy(weightsDims[1:], outputDims)
+		weightsDims := make([]int, 0, 2+len(outputDims))
+		if isEnsemble {
+			weightsDims = append(weightsDims, c.ensembleSize)
+		}
+		weightsDims = append(weightsDims, inputLastDimension)
+		weightsDims = append(weightsDims, outputDims...)
 		weightsVar := layerCtx.VariableWithShape("weights", shapes.Make(dtype, weightsDims...))
 		if c.regularizer != nil {
 			// Only for the weights, not for the bias.
@@ -281,10 +338,84 @@ func (c *Config) Done() *Node {
 		weights := weightsVar.ValueGraph(g)
 		var biasNode *Node
 		if c.useBias {
-			biasVar := layerCtx.VariableWithShape("biases", shapes.Make(dtype, outputDims...))
+			biasDims := make([]int, 0, 1+len(outputDims))
+			if isEnsemble {
+				biasDims = append(biasDims, c.ensembleSize)
+			}
+			biasDims = append(biasDims, outputDims...)
+			biasVar := layerCtx.VariableWithShape("biases", shapes.Make(dtype, biasDims...))
 			biasNode = biasVar.ValueGraph(g)
 		}
-		x = nn.Dense(x, weights, biasNode)
+
+		if !isEnsemble {
+			x = nn.Dense(x, weights, biasNode)
+		} else if c.ensembleAxis >= 0 {
+			if ii == 0 {
+				x = Dot(x, weights).General([]int{x.Rank() - 1}, []int{c.ensembleAxis}, []int{1}, []int{0})
+			} else {
+				x = Dot(x, weights).General([]int{x.Rank() - 1}, []int{0}, []int{1}, []int{0})
+			}
+			if biasNode != nil {
+				batchRank := x.Rank() - 1 - len(outputDims)
+				axesToInsert := make([]int, batchRank)
+				for j := 0; j < batchRank; j++ {
+					axesToInsert[j] = 1 + j
+				}
+				x = Add(x, ExpandAxes(biasNode, axesToInsert...))
+			}
+		} else {
+			switch ii {
+			case 0:
+				x = Dot(x, weights).General([]int{x.Rank() - 1}, nil, []int{1}, nil)
+				if biasNode != nil {
+					x = Add(x, ExpandLeftToRank(biasNode, x.Rank()))
+				}
+			case 1:
+				x = Dot(x, weights).General([]int{x.Rank() - 1}, []int{x.Rank() - 2}, []int{1}, []int{0})
+				if biasNode != nil {
+					batchRank := x.Rank() - 1 - len(outputDims)
+					axesToInsert := make([]int, batchRank)
+					for j := 0; j < batchRank; j++ {
+						axesToInsert[j] = 1 + j
+					}
+					x = Add(x, ExpandAxes(biasNode, axesToInsert...))
+				}
+			default:
+				x = Dot(x, weights).General([]int{x.Rank() - 1}, []int{0}, []int{1}, []int{0})
+				if biasNode != nil {
+					batchRank := x.Rank() - 1 - len(outputDims)
+					axesToInsert := make([]int, batchRank)
+					for j := 0; j < batchRank; j++ {
+						axesToInsert[j] = 1 + j
+					}
+					x = Add(x, ExpandAxes(biasNode, axesToInsert...))
+				}
+			}
+		}
 	}
+
+	if isEnsemble {
+		needsTranspose := false
+		pos := x.Rank() - 1 - len(c.outputDimensions)
+		if c.ensembleAxis >= 0 {
+			pos = c.ensembleAxis
+			needsTranspose = true
+		} else if c.numHiddenLayers > 0 {
+			needsTranspose = true
+		}
+
+		if needsTranspose {
+			perm := make([]int, x.Rank())
+			for j := 0; j < pos; j++ {
+				perm[j] = j + 1
+			}
+			perm[pos] = 0
+			for j := pos + 1; j < x.Rank(); j++ {
+				perm[j] = j
+			}
+			x = TransposeAllAxes(x, perm...)
+		}
+	}
+
 	return x
 }