+++ docs/CHANGELOG.md
@@ -37,6 +37,8 @@
 - Package `attention/pos`:
   - Split `Encoder` API into `QKEncoder` and `PreEncoder`, to support different types of positional encoders.
   - Added `WithSlidingWindow` to support "sliding attention" (the slow way).
+- Package `layers/fnn`:
+  - Added `WithEnsembleSize` and `WithEnsembleAxis` methods for configuring parallel independent executions via ensembles.
 - Package `transfromers`:
   - Updates and fixes to the API; Added methods to build partial models: `AllLayers`, `ForwardLayer`, `LogitsFromEmbeddings`, `EmbedTokesn`, etc.
   - Updated positional-encoder support.